refactor: check -> ticket
